added IndividualConstraint.js and refactored Layer.js. Removed LayerHelpers.js
